test: AdminControllerTest 내부 테스트들 수정
